Split sqlite store core helpers
